Making the return types more intelligable.
