+++ src/main/java/com/google/maps/internal/OkHttpPendingResult.java
@@ -20,7 +20,7 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonSyntaxException;
 import com.google.maps.GeolocationApi;
-import com.google.maps.ImageResponse;
+import com.google.maps.ImageResult;
 import com.google.maps.PendingResult;
 import com.google.maps.errors.ApiException;
 import com.google.maps.model.AddressComponentType;
@@ -232,9 +232,9 @@ public class OkHttpPendingResult<T, R extends ApiResponse<T>>
 
     if (contentType != null
         && contentType.startsWith("image")
-        && responseClass == ImageResponse.class
+        && responseClass == ImageResult.Response.class
         && response.code() == 200) {
-      ImageResponse.Result result = new ImageResponse.Result(contentType, bytes);
+      ImageResult result = new ImageResult(contentType, bytes);
       return (T) result;
     }
